Node joining the cluster process. Things still to be improved.
